fix: create authentication folder to store athentications process and separate login logic from the route
