+++ server/App.js
@@ -7,7 +7,7 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-var loginRouter = require('./routes/login');
+var loginRouter = require('./authentication/authRouter')
 var mailRouter = require('./routes/mail')
 
 const middleware = require('./utils/middleware');
