+++ server/authentication/authRouter.js
@@ -1,0 +1,6 @@
+const { loginController } = require("./authController");
+
+const loginRouter = require("express").Router();
+
+loginRouter.post('/', loginController)
+module.exports = loginRouter
